styling buttons and right side of nav bar
