+++ src/app/navbar/navbar.css
@@ -2,33 +2,51 @@
     background-color: var(--dark-accent-color);
     color: #ffffff;
     width: 100%;
-    height: 55px;
+    height: 4rem;
+    display:flex;
+    align-items: center;
+    padding: 1rem;
 }
-
+.nav-left{
+    display:flex;
+    align-items: center;
+    font-weight: 600;
+}
+.icon{
+    height: 1.5rem;
+}
 ul {
     list-style-type: none;
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
+    display: flex;
+    align-items: center;
+    margin-right: 0.5rem;
 }
 
 li {
-    float: left;
-    padding-left: 20px;
+    margin: 0 0.5rem;
+    cursor: pointer;
+    white-space: nowrap;
     display: flex;
-    justify-content: center;
-    align-content: center;
-    flex-direction:column;
-    height: 50px;
 }
 
-.right {
-    float: right !important;
-    padding-right: 20px;
+a{
+    text-decoration: none;
+    color: #ffffff;
+    &:hover{
+        color: #afddfc ;
+    }
 }
 
-.centre {
-    position: absolute;
-    left: 50%;
-    transform: translate(-50%);
+.nav-button{
+    background-color: #afddfc;
+    color: var(--dark-accent-color);
+    border: none;
+    cursor: pointer;
+    font-size: 1rem;
+    font-weight: 600;
+    padding: 0.35rem 1rem;
+    &:hover{
+        color: #afddfc ;
+        border:#afddfc solid 2px;
+    }
 }
